back/requirements.txt
fastapi[standard]
pypdf
PyMuPDF
Pillow
groq
elasticsearch
# sentence-transformers
langchain-text-splitters